meiduo/requirements.txt
itsdangerous==1.1.0
djangorestframework_jwt==1.11.0
celery==5.0.5
requests==2.25.1
django_redis==4.12.1
drf_extensions==0.7.0
redis==3.5.3
Django==3.1.7
PyMySQL==1.0.2
djangorestframework==3.12.2
